Move validate_application_error_event_count into validators directory
