+++ tests/agent_features/test_notice_error.py
@@ -19,9 +19,11 @@ from testing_support.fixtures import (
     error_is_saved,
     override_application_settings,
     reset_core_stats_engine,
-    validate_application_error_event_count,
     validate_transaction_error_event_count,
     validate_transaction_error_trace_count,
+)
+from testing_support.validators.validate_application_error_event_count import (
+    validate_application_error_event_count,
 )
 from testing_support.validators.validate_application_error_trace_count import (
     validate_application_error_trace_count,
